+++ NEWS.md
@@ -1,3 +1,15 @@
+# lmtp 1.5.2
+
+### New Features
+
+### Bug Fixes
+
+### General
+
+- The default for the `mtp` argument has been changed from `FALSE` to `TRUE` (see issue \#170).
+- A warning message is now printed in `lmtp_contrast()` that p-values aren't adjusted for multiple comparisons (see issue \#172).
+- A warning message is now printed in `lmtp_contrast()` if `ref` is a constant.
+
 # lmtp 1.5.1
 
 ### New Features